+++ source/private/Connect-TcpWithTimeout.ps1
@@ -51,12 +51,8 @@ function Connect-TcpWithTimeout {
             Write-Debug "[$fn] Failed to dispose TcpClient after connection error: $($_.Exception.GetType().FullName)"
         }
 
-        $errorToThrow = $_.Exception
-        $throwOriginal = $true
-        if ($errorToThrow -is [System.AggregateException] -and $errorToThrow.InnerException) {
-            $errorToThrow = $errorToThrow.InnerException
-            $throwOriginal = $false
-        }
+        $errorToThrow = Resolve-TlsException -Exception $_.Exception
+        $throwOriginal = [object]::ReferenceEquals($errorToThrow, $_.Exception)
 
         $null = Add-TlsErrorContext `
             -Exception $errorToThrow `